1. redis 存储Long数据反序列化成了Inter，暂未解决 2. redis 存储token乱码已解 -》 redisConfig 3. jwt 登陆已走通 4. todo：jwt success response

+++ src/main/java/com/example/securitydemo/mybatis/dao/UserDao.java
@@ -8,13 +8,13 @@ import org.springframework.stereotype.Repository;
 @Mapper
 @Repository
 public interface UserDao {
-    User queryUserById(Long id);
+    User queryUserById(int id);
 
-    MyUserDetails queryUserDetailsById(Long id);
+    MyUserDetails queryUserDetailsById(int id);
 
     MyUserDetails queryUserDetailsByPhone(String phone);
 
-    Long queryIdByPhone(String phone);
+    int queryIdByPhone(String phone);
 
     int insertUser(User user);
 }
